refactor: home page tests
